Improve introductionary documentation
De-emphasize working directly with the Value type as this is not the
common (or recommended) way to use the library, but still include
documentation on how to use it when it's called for.

Since the documentation on using generics is documented in the FromJSON
and ToJSON classes, refer to that documentation instead of only
documenting one method (which is not the preferred method anymore with
default signatures I believe).