Fix .:? behavior for #83
It makes (.:?) return Just value if the key is present for values of type Maybe a.
Earlier {"key": null} .:? "key" would return Nothing although the key is there. It was not possible to distinguish the cases when key is absent and when value is null.
Now it treats Maybe same way as other types and returns Just Nothing.