Avoid (10 ^ e) on fromScientific
On [1], @basvandijk noted that fromScientific used 10 ^ e.  This
operation could lead to a DOS due to a large exponent.

This commit changes the implementation to always use
formatScientificBuilder, even when base10Exponent s >= 0.  The code from
the scientific package already takes care when pretty printing huge
numbers.

Note that using just scientificBuilder wouldn't be enough since we'd
change the formatting of large integers to that of a large Double.  This
commit does not change the output format at all, and all tests are
green.

[1] https://github.com/bos/aeson/issues/246#issuecomment-93599848