Add an attoparsec based UTC time decoder (#197)
Replace the old 'parseTime' based UTCTime parser with an attoparsec
based parser.  This is roughly 5x faster than the old one.

It also allows decoding some timestamp formats that the old version
rejected.  The new version still requires all timestamps to specify a
timezone when decoding to a UTCTime type.

Also add a few unit tests for both what should be accepted and also
for things that should be rejected.

A lot of this code was bummed from the postgresql-simple library -
thus credits to @lpsmith too!