Fix utc offset parsing
As I understand ISO 8601:2004,  ±HHMM is not a valid syntax for specifying
a UTC offset in the context we are using it.   Only ±HH and ±HH:MM are
valid in the extended ISO 8601 formats.

I probably would weakly prefer to not accept ±HHMM, however,  because the
the ToJSON ZonedTime instance has been emitting this syntax since aeson
version 0.6.1.0 released on 2012-12-14,   our hands may well be tied on
this matter.

I can only hope that the older FromJSON instances will also accept ±HH and
±HH:MM,  so that we can fix the emitted syntax with minimal disruption,
but knowing the time library, this probably is not the case.