Split round trip test into round trip encoding and round trip toJSON
Previously, the test was converting an object to json using `toJSON`
and then encoding *that* (which always had a type of Value), so
the test was not catching cases where there was a bug in the
encoding function but not the toJSON function. This change makes
it so that both methods are always tested independently.