Add JSONTestSuite support as unit tests
That repo currently contains 335 tests, of which 299 pass.

Of the 36 failing tests, some are true bugs, and of those true bugs,
some are probably worth fixing, but I haven't triaged any yet.

NOTE: to run the test suite, you now need either a copy of the
JSONTestSuite repo in the tests directory, or a symlink to a copy
that lives somewhere else.