Re-export some TH type names in Data.Aeson.TH
When writing functions that call `deriveJSON` with different options,
type names such as `Name` or `Q` are required for the function's type
signature. As they are not exported by this package, the user must
separately import `Language.Haskell.TH` and add `template-haskell` to
the project `build-depends` section. Creating wrappers of `deriveJSON`
is inconvenient because of this.

This commit attempts to relieve the pain in a minimally invasive way.

As far as I know, this is a mostly backwards-compatible change.