Add option to reject unknown fields (#207)
This commit adds rejectUnknownFields as a Option that affects parsing.
Parsing will succeed only if all of the fields in the raw JSON value
are present in the record to which it will be parsed.

It supports both Generics and TH, though the TH implementation has less
detailed messaging. This seems to be a pattern througout though, so it
should be consistent with other parsing errors.