Improve `Arbitrary Value` instance.
Fixes #980. As discussed there, simply changing `n <= 0` to `n <= 1`
gives good generation of strings, bools, numbers and null. Turns out
it also forbids `[]` and `{}` from being generated, so I added those
explicitly to that case.

My understanding is this requires a major version bump, as a change in
behavior of an exported function.