Add Data.Aeson.Decoding working on token stream
There are two motivation points:

1. it's faster
2. it's more expressive

The first point is actually accidental, but nevertheless very nice to
have. In benchmarks decoding strict bytestring through Tokens
seems to be 30% or so faster for examples we have:

```
  Examples
    decode
      github-issues
        lazy:   OK (0.14s)
          2.03 ms ± 168 μs
        strict: OK (0.25s)
          1.94 ms ± 180 μs
        tokens: OK (0.34s)
          1.27 ms ±  61 μs

  Twitter
    decode
      Generic
        direct
          twitter100:  OK (1.30s)
            2.43 ms ± 161 μs
          jp100:       OK (0.25s)
            2.09 ms ± 126 μs
          twitter100t: OK (0.33s)
            1.34 ms ± 121 μs
          jp100t:      OK (0.17s)
            1.37 ms ±  86 μs
      Manual
        direct
          twitter100:  OK (0.21s)
            1.76 ms ±  86 μs
          jp100:       OK (0.30s)
            2.29 ms ± 207 μs
          twitter100t: OK (0.74s)
            1.38 ms ±  33 μs
          jp100t:      OK (0.20s)
            1.61 ms ± 125 μs
      TH
        direct
          twitter100:  OK (0.27s)
            2.08 ms ± 180 μs
          jp100:       OK (0.54s)
            2.17 ms ± 149 μs
          twitter100t: OK (0.36s)
            1.38 ms ±  50 μs
          jp100t:      OK (0.39s)
            1.48 ms ±  49 μs
```

The second point is also quite important.
If some users really need to know the order of key-value pairs
in a JSON object, or care whether number was integral, decimal or
scientific; now they can differentiate.

Though they cannot use `FromJSON` machinery, this is a start.

E.g. issues like *Should we make a clear value-level distinction between
floating and integral numbers?* https://github.com/haskell/aeson/issues/546
(For `Value` it would be bad, for `Tokens` it's fine).